+++ constraints/anti_infective/AntiInfectiveBundle.cpp
@@ -1,1 +1,4 @@
 #include "AntiInfectiveBundle.hpp"
+#include "../../constraint_core/ConstraintRegistry.hpp"
+
+REGISTER_CONSTRAINT_BUNDLE("AntiInfective", AntiInfectiveBundle)