+++ constraints/cardiac/CardiacBundle.cpp
@@ -1,1 +1,4 @@
 #include "CardiacBundle.hpp"
+#include "../../constraint_core/ConstraintRegistry.hpp"
+
+REGISTER_CONSTRAINT_BUNDLE("Cardiac", CardiacBundle)