+++ constraints/cns/CNSBundle.hpp
@@ -2,7 +2,6 @@
 #define CURAFRAME_CNS_BUNDLE_HPP
 
 #include "../../constraint_core/ConstraintBundle.hpp"
-#include "../../constraint_core/ConstraintRegistry.hpp"
 #include <sstream>
 
 class CNSBundle : public ConstraintBundle {
@@ -58,7 +57,5 @@ public:
     std::vector<std::string> falsification_flags() const override { return flags; }
     std::string narrative_summary() const override { return summary; }
 };
-
-REGISTER_CONSTRAINT_BUNDLE("CNS", CNSBundle)
 
 #endif // CURAFRAME_CNS_BUNDLE_HPP