+++ constraints/metabolic/MetabolicBundle.cpp
@@ -1,1 +1,4 @@
 #include "MetabolicBundle.hpp"
+#include "../../constraint_core/ConstraintRegistry.hpp"
+
+REGISTER_CONSTRAINT_BUNDLE("Metabolic", MetabolicBundle)