+++ constraints/metabolic/MetabolicBundle.hpp
@@ -2,7 +2,6 @@
 #define CURAFRAME_METABOLIC_BUNDLE_HPP
 
 #include "../../constraint_core/ConstraintBundle.hpp"
-#include "../../constraint_core/ConstraintRegistry.hpp"
 #include <sstream>
 
 class MetabolicBundle : public ConstraintBundle {
@@ -74,7 +73,5 @@ public:
     std::vector<std::string> falsification_flags() const override { return flags; }
     std::string narrative_summary() const override { return summary; }
 };
-
-REGISTER_CONSTRAINT_BUNDLE("Metabolic", MetabolicBundle)
 
 #endif // CURAFRAME_METABOLIC_BUNDLE_HPP