+++ constraints/oncology/OncologyBundle.cpp
@@ -1,1 +1,4 @@
 #include "OncologyBundle.hpp"
+#include "../../constraint_core/ConstraintRegistry.hpp"
+
+REGISTER_CONSTRAINT_BUNDLE("Oncology", OncologyBundle)