+++ constraints/renal/RenalBundle.cpp
@@ -1,1 +1,4 @@
 #include "RenalBundle.hpp"
+#include "../../constraint_core/ConstraintRegistry.hpp"
+
+REGISTER_CONSTRAINT_BUNDLE("Renal", RenalBundle)