+++ constraints/systemic_exposure/SystemicExposureBundle.cpp
@@ -1,1 +1,4 @@
 #include "SystemicExposureBundle.hpp"
+#include "../../constraint_core/ConstraintRegistry.hpp"
+
+REGISTER_CONSTRAINT_BUNDLE("SystemicExposure", SystemicExposureBundle)